Made visuals_core_2 with working state
Made a second visuals core, with the old being renamed to proto. So far, this one has the ability to change the state of a button, but the infrastructure exists to build much more.

To-do: Currently I still need to implement the basic Image component to use for backgrounds and such. Farther in the future, I want to get full animated sprite sheets to work, but that is a much larger project.